ai/requirements.txt
google-generativeai==0.3.2
numpy==1.24.3
pandas==2.0.3
flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0 